Add files via upload
Detailed Project Proposal
We will create a simple app for divination. In this app, we will use the Six Lines (Liu Yao) method for divination. Six Lines is an ancient Chinese divination method that consists of six lines, which can be either yin(————  ————) (or yang(——————————). These six lines are divided into two groups, with each group corresponding to one of the trigrams from the Eight Trigrams (Ba Gua).

During the coding phase, the user will be prompted to input three three-digit numbers. After the input is completed, the first and second three-digit numbers will be divided by 8, and the third number will be divided by 6. The remainder of each division will then be taken.
In traditional Six Lines (Liu Yao) divination, the numbers 1 to 8 correspond to one of the trigrams (each composed of three lines). The remainder of the first number will determine the upper half of the Six Lines hexagram, while the remainder of the second number will determine the lower half.
The remainder of the third number will serve as the "changing line," which modifies the hexagram formed by the first two remainders. For example, if the remainder of the third number is 1, the bottom line of the hexagram will change (if it is a yin line, it will become a yang line, and vice versa). If the third remainder is 2, the second line from the bottom will change, and so on.
In the subsequent coding and app design, we will generate text based on the above results to provide divination interpretations for the user.